refactor and organize shell crate code
